rename github -> github-tags
