+++ lib/datasource/go/index.spec.ts
@@ -1,5 +1,5 @@
 import _got from '../../util/got';
-import * as _github from '../github';
+import * as _github from '../github-tags';
 import * as go from '.';
 import { mocked, partial } from '../../../test/util';
 import { ReleaseResult } from '..';
