+++ lib/datasource/go/index.ts
@@ -1,6 +1,6 @@
 import { logger } from '../../logger';
 import got from '../../util/got';
-import * as github from '../github';
+import * as github from '../github-tags';
 import { DigestConfig, PkgReleaseConfig, ReleaseResult } from '../common';
 import { regEx } from '../../util/regex';
 import {
